fix: move life cycle functions into the single file
